# lib/api-zod/src/generated/types/userUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Staff Portal API
 * OpenAPI spec version: 0.1.0
 */

export interface UserUpdate {
  /**
     * @minLength 1
     * @maxLength 150
     */
  full_name?: string;
  /**
     * @minLength 1
     * @maxLength 50
     */
  username?: string;
  email?: string;
  /** @minLength 6 */
  password?: string;
  /** @nullable */
  department_id?: number | null;
  role_ids?: number[];
}